+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [0.1.0] - 19/11/2025
+### Added
+- Added modify for workshops.
 
 ### Added
  - US for modify supply batch
@@ -10,6 +12,7 @@
 - Fixed search bar in workshop list.
 - Fixed the product error message
 - Error now properly appearing when registering new beneficiary
+- Error in descriptcion in view and modify a workshop
 - Fixed error views in product batches list, detail, register and modify
 - Fixed product fetch in register Product Batch view